requirements.txt
# Core data and computation
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Additional utilities
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0

# Trading and broker integration
alpaca-trade-api>=3.0.0
pytz>=2023.3
